write data to shared application group

+++ bither-ios/Utils/GroupFileUtil.h
@@ -38,7 +38,7 @@ typedef enum {
 
 +(void)setDefaultBitcoinUnit:(GroupBitcoinUnit)unit;
 
-+(void)setTotalBalanceWithHDM:(long long)hdm hot:(long long)hot andCold:(long long)cold;
++(void)setTotalBalanceWithHDM:(int64_t)hdm hot:(int64_t)hot andCold:(int64_t)cold;
 
 +(NSDictionary*)totalBalance;
